# Conflicts:
#	linkwe-api/src/main/java/com/linkwechat/controller/WeIndexController.java
#	linkwe-common/src/main/java/com/linkwechat/common/config/LinkWeChatConfig.java
#	linkwe-service/src/main/java/com/linkwechat/domain/customer/vo/WeCustomerChannelCountVo.java
#	linkwe-service/src/main/java/com/linkwechat/service/IWeCustomerService.java
#	linkwe-service/src/main/java/com/linkwechat/service/IWeGroupService.java
#	linkwe-service/src/main/java/com/linkwechat/service/impl/WeCustomerServiceImpl.java
#	linkwe-service/src/main/java/com/linkwechat/service/impl/WeGroupCodeServiceImpl.java
#	linkwe-service/src/main/java/com/linkwechat/service/impl/WeGroupServiceImpl.java
#	linkwe-service/src/main/resources/mapper/wecom/WeCustomerMapper.xml
#	sql/update.md

+++ linkwe-service/src/main/java/com/linkwechat/domain/kf/WeKfUser.java
@@ -20,7 +20,7 @@ public class WeKfUser {
     private String userName;
 
     @ApiModelProperty("部门id")
-    private Integer departmentId;
+    private Long departmentId;
 
     @ApiModelProperty("部门名称")
     private String deptName;
